Parsing (#8)
Overview
---------------
Addresses issue #8 by implementing a pair of read/write functions that will either return/accept a numpy array or a pandas DataFrame. Headers are a mandatory input for these functions but a default set of values appropriate for 5 particle coincidence is given as a default input. The write function only checks if a numpy array is passed and assumes that if a numpy array wasn't passed then a pandas DataFrame was passed. Potentially this could result in future issues if other data structures are passed into this function as the error message might not end up being helpful.


Commit History
-------------------------

* initial implementation of parsing module to read and write ' ' seperated momenta data files as well as tests for those read/write functions

* fixing bugs in output txt file writting and typos in doc strings

* increasing the number of integers used to check the read/write functionality

* fixing linter errors

* fixing typos in the doc strings

* fixing coverage issue and logic bug in write momentum file

* allowing the headers to be modifiable to allow for different momentum datasets with different numbers of columns to be read by the same function

* fixing syntax error and linter error